research_pipeline copy/requirements.txt
arxiv==2.0.0
openai>=1.0.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
matplotlib==3.8.2
seaborn==0.13.2
pandas==2.2.0
numpy==1.26.3
tenacity==8.2.3
aiohttp==3.9.3
